+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2
+
+- Handling multiple query parameter values.
+- Option to sort query parameter values.
+
 ## 1.0.1
 
 - Handle URLs without hosts.